Add basic receiver routine
